requirements.txt
scikit-learn
pandas
plotly
python-dotenv==1.0.0
cosmpy>=0.9.2
nillion-python-helpers
py-nillion-client==0.5.0
nada-dsl==0.5.0
nada-numpy==0.4.0
nada-ai==0.4.0